% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendDatabase.R
\name{addDatabaseIdToTables}
\alias{addDatabaseIdToTables}
\title{Add databaseId column to tables in results *.zip files}
\usage{
addDatabaseIdToTables(
  tableName,
  originalZipFileName,
  newZipFileName,
  databaseId,
  tempFolder = tempdir()
)
}
\arguments{
\item{originalZipFileName}{Name of single or vector of multiple zip files.}

\item{databaseId}{Name or single or vector of multiple databaseId to add}

\item{tempFolder}{Directory which to use to unpack files}

\item{newZipFilaName}{Name of single or vector of multiple zip files}
}
\description{
Adds databaseId column to tables that are missing such a column in results *.zip files
}
