% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assessment.R
\name{assessPhenotypes}
\alias{assessPhenotypes}
\title{Assess propensity models}
\usage{
assessPhenotypes(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  tablePrefix = "legend_t2dm",
  indicationId = "class",
  oracleTempSchema,
  outputFolder,
  sampleSize = 1e+05,
  minCellCount = 5,
  databaseId,
  databaseName = databaseId,
  databaseDescription = databaseId,
  createExposureCohorts = TRUE,
  runExposureCohortDiagnostics = TRUE,
  createOutcomeCohorts = TRUE,
  runOutcomeCohortDiagnostics = TRUE,
  filterExposureCohorts = NULL,
  filterOutcomeCohorts = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in
the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your vocabulary tables in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}

\item{tablePrefix}{A prefix to be used for all table names created for this study.}

\item{indicationId}{A string denoting the indicationId for which the exposure cohorts
should be created; should be 'class' or 'drug'}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward
slashes (/)}

\item{sampleSize}{What is the maximum sample size across exposure cohorts?}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{Full name for the database.}

\item{databaseDescription}{Brief description of population in database}

\item{createExposureCohorts}{Boolean: execute exposure cohort instantiation? \code{FALSE} will
attempt to re-use the cohort table in \code{cohortDatabaseSchema}}

\item{runExposureCohortDiagnostics}{Boolean: execute cohort diagnostics on exposure cohorts?}

\item{createOutcomeCohorts}{Boolean: execute outcome cohort instantiation? \code{FALSE} will
attempt to re-use the cohort table in \code{cohortDatabaseSchema}}

\item{runOutcomeCohortDiagnostics}{Boolean: execute cohort diagnostics on outcome cohorts?}

\item{filterExposureCohorts}{Optional subset of exposure cohorts to use; \code{NULL} implies all.}

\item{filterOutcomeCohorts}{Options subset of outcome cohorts to use; \code{NULL} implies all.}
}
\description{
Assess propensity models
}
\details{
This function will sample the exposure cohorts, and fit propensity models to identify issues.
Assumes the exposure and outcome cohorts have already been created.
}
