% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assessment.R
\name{assessPropensityModels}
\alias{assessPropensityModels}
\title{Assess propensity models}
\usage{
assessPropensityModels(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  tablePrefix = "legend_t2dm",
  indicationId = "class",
  oracleTempSchema,
  outputFolder,
  minCohortSize = 1000,
  sampleSize = 1000,
  maxCores = 4,
  databaseId,
  preferenceScoreBounds = c(0.3, 0.7),
  forceNewCmDataObjects = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in
the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}

\item{tablePrefix}{A prefix to be used for all table names created for this study.}

\item{indicationId}{A string denoting the indicationId for which the exposure cohorts
should be created; should be 'class' or 'drug'}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward
slashes (/)}

\item{minCohortSize}{Minimum number of people that have to be in each cohort to keep a pair of
cohorts.}

\item{sampleSize}{What is the maximum sample size for each exposure cohort?}

\item{maxCores}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{preferenceScoreBounds}{Preference score bounds to use when reporting proportion of subjects in
empirical clinical equipoise.}

\item{forceNewCmDataObjects}{Force recreation of \code{CohortMethod} data objects?}
}
\description{
Assess propensity models
}
\details{
This function will sample the exposure cohorts, and fit propensity models to identify issues.
Assumes the exposure and outcome cohorts have already been created.
}
