% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateBalance.R
\name{computeCovariateBalance}
\alias{computeCovariateBalance}
\title{Compute covariate balance for each comparison}
\usage{
computeCovariateBalance(indicationId = "legendt2dm", outputFolder, maxCores)
}
\arguments{
\item{indicationId}{A string denoting the indicationId for which the covariate balance should be
computed.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}

\item{maxCores}{How many parallel cores should be used? If more cores are made available this
can speed up the analyses.}
}
\description{
Compute covariate balance for each comparison
}
\details{
Compute covariate balance for each comparison. We do not compute balance for all covariates for every TCO,
as this would be prohibitively large. Instead, we compute balance considering the entire target
and comparator cohort, so not removing subjects with prior outcomes. We do compute covariate balance
for every TCO for a small subset of covariates of interest (those used in the typical table 1).
Stores the results in a subfolder called 'balance'.
}
