% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendDatabase.R
\name{createDataModelSqlFile}
\alias{createDataModelSqlFile}
\title{Create a SQL file to construct data model tables on a database server.}
\usage{
createDataModelSqlFile(specifications, fileName, tab = "    ")
}
\arguments{
\item{specifications}{Specifications data table}

\item{fileName}{Output name for SQL file}

\item{tab}{Tab characters to use}
}
\description{
Create a SQL file to construct data model tables on a database server.
}
\details{
Only PostgreSQL servers are supported.
}
