% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{doMetaAnalysis}
\alias{doMetaAnalysis}
\title{Perform random-effects meta-analysis on CES results on data server}
\usage{
doMetaAnalysis(
  connectionDetails,
  indicationId = "class",
  resultsDatabaseSchema,
  maExportFolder,
  maName = "Meta-analysis",
  diagnosticsFilter = NULL,
  maxCores,
  cacheFileName = NULL
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector details to public LEGEND-T2DM db server}

\item{resultsDatabaseSchema}{Schema on the postgres server where the tables have been created.}

\item{maExportFolder}{A local folder where the meta-anlysis results will be written.}

\item{diagnosticsFilter}{Table of which target-comparator-outcome-analysis-database tuples pass diagnostics;
can be NULL}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{cacheFileName}{Use (if exists) or cache database results in file}
}
\description{
Perform random-effects meta-analysis on CES results on data server
}
\details{
This function executes a meta-analysis across data source results.
}
