% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI LEGEND study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema,
  cohortDatabaseSchema,
  outputFolder,
  indicationId = "class",
  tablePrefix = "legendt2dm",
  databaseId = "Unknown",
  databaseName = "Unknown",
  databaseDescription = "Unknown",
  minCohortSize = 1000,
  minCellCount = 5,
  imputeExposureLengthWhenMissing = FALSE,
  studyEndDate = "",
  createExposureCohorts = TRUE,
  createOutcomeCohorts = TRUE,
  fetchAllDataFromServer = TRUE,
  synthesizePositiveControls = FALSE,
  generateAllCohortMethodDataObjects = TRUE,
  runCohortMethod = TRUE,
  runSections = c(1:6),
  computeCovariateBalance = TRUE,
  exportToCsv = TRUE,
  filterExposureCohorts = NULL,
  filterOutcomeCohorts = NULL,
  maxCores = 4
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your vocabulary tables in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the
user has write priviliges for storing temporary tables.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You
will need to have write priviliges in this schema. Note
that for SQL Server, this should include both the
database and schema name, for example 'cdm_data.dbo'.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{indicationId}{A string denoting the indicationId.}

\item{tablePrefix}{A prefix to be used for all table names created for this
study.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{databaseName}{The full name of the database (e.g. 'Medicare Claims
Synthetic Public Use Files (SynPUFs)').}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{minCellCount}{The minimum cell count for fields contains person counts
or fractions when exporting to CSV.}

\item{imputeExposureLengthWhenMissing}{For OptumEHR: impute length of drug exposures when the
length is missing?}

\item{studyEndDate}{Optional study end date for EHRs}

\item{createExposureCohorts}{Create the tables with the exposure cohorts?}

\item{createOutcomeCohorts}{Create the tables with the outcome cohorts?}

\item{fetchAllDataFromServer}{Fetch all relevant data from the server?}

\item{synthesizePositiveControls}{Inject signals to create synthetic controls?}

\item{generateAllCohortMethodDataObjects}{Create the cohortMethodData objects from the fetched
data and injected signals?}

\item{runCohortMethod}{Run the CohortMethod package to produce the outcome
models?}

\item{runSections}{Run specific sections through CohortMethod}

\item{computeCovariateBalance}{Report covariate balance statistics across comparisons?}

\item{exportToCsv}{Export all results to CSV files?}

\item{filterExposureCohorts}{Optional subset of exposure cohorts to use; \code{NULL} implies all.}

\item{filterOutcomeCohorts}{Options subset of outcome cohorts to use; \code{NULL} implies all.}

\item{maxCores}{How many parallel cores should be used? If more cores
are made available this can speed up the analyses.}
}
\description{
Execute OHDSI LEGEND study
}
\details{
This function executes the OHDSI LEGEND study.
}
