% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCmDataObjectBuilding.R
\name{fetchAllDataFromServer}
\alias{fetchAllDataFromServer}
\title{Fetch all data on the cohorts for analysis}
\usage{
fetchAllDataFromServer(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  oracleTempSchema,
  indicationId = "class",
  tablePrefix = "legendt2dm",
  useSample = FALSE,
  forceNewObjects = FALSE,
  studyEndDate = "",
  outputFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in
the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{indicationId}{A string denoting the indicationId.}

\item{tablePrefix}{A prefix to be used for all table names created for this study.}

\item{useSample}{Use the sampled cohort table instead of the main cohort table (for PS
model feasibility).}

\item{forceNewObjects}{Force recreation of \code{CohortMethod} data objects?}

\item{studyEndDate}{Optionally specify a study end date}

\item{outputFolder}{Schema name where intermediate data can be stored. You will need to
have write priviliges in this schema. Note that for SQL Server, this
should include both the database and schema name, for example
'cdm_data.dbo'.}
}
\description{
Fetch all data on the cohorts for analysis
}
\details{
This function will create covariates and fetch outcomes and person information from the server.
}
