% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterByExposureCohortsSize.R
\name{filterByExposureCohortsSize}
\alias{filterByExposureCohortsSize}
\title{Filter exposure pairs by size of the cohorts}
\usage{
filterByExposureCohortsSize(
  outputFolder,
  indicationId,
  minCohortSize,
  summaryFile = NULL
)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}

\item{indicationId}{A string denoting the indicationId.}

\item{minCohortSize}{Minimum number of people that have to be in each cohort to keep a pair of
cohorts.}

\item{summaryFile}{File name of the csv that contains paired exposure summary;
if NULL, then default to "pairedExposureSummary.csv"}
}
\description{
Filter exposure pairs by size of the cohorts
}
