% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCmDataObjectBuilding.R
\name{generateAllCohortMethodDataObjects}
\alias{generateAllCohortMethodDataObjects}
\title{Construct all cohortMethodData object}
\usage{
generateAllCohortMethodDataObjects(
  outputFolder,
  indicationId = "legendt2dm",
  useSample = FALSE,
  restrictToOt1 = FALSE,
  maxCores = 4
)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}

\item{indicationId}{A string denoting the indicationId.}

\item{useSample}{Use the sampled cohort table instead of the main cohort table (for PS model
feasibility).}

\item{restrictToOt1}{Limit \code{CohortMethod} data objects to only comparisons that use On-Treatment-1?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available this
can speed up the analyses.}
}
\description{
Construct all cohortMethodData object
}
\details{
This function constructs all cohortMethodData objects using the data fetched earlier using the
\code{\link{fetchAllDataFromServer}} function.
}
