% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareStagedExecution.R
\name{prepareStagedExecution}
\alias{prepareStagedExecution}
\title{Utility function to split up CES execution, with only a subset of TC pairs in each run}
\usage{
prepareStagedExecution(
  originalOutputFolder,
  outputFolderHeader = originalOutputFolder,
  indicationId = "drug",
  stages = 10
)
}
\arguments{
\item{originalOutputFolder}{Original output folder used when creating cohorts}

\item{outputFolderHeader}{The staged execution output folders will start with this header name;
default is the same as `originalOutputFolder`}

\item{indicationId}{Which sub-study to run?}

\item{stages}{Number of stages for execution?}
}
\description{
Utility function to split up CES execution, with only a subset of TC pairs in each run
}
\details{
This function prepares staged CES execution by copying/slicing necessary files.
But note, this does NOT run the study; this is only a utility function that allows staged execution.
}
