% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printCohortDefinition}
\alias{printCohortDefinition}
\title{Print cohort definition from file name}
\usage{
printCohortDefinition(info)
}
\arguments{
\item{info}{List with two entries: \code{name} (text name for printing) and
\code{jsonFileName} (JSON file name)}
}
\description{
Outputs a cohort definition into human-readable \code{markdown}
}
