% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printCohortDefinitionFromNameAndJson}
\alias{printCohortDefinitionFromNameAndJson}
\title{Print cohort definition with specified name}
\usage{
printCohortDefinitionFromNameAndJson(
  name,
  json = NULL,
  obj = NULL,
  withConcepts = TRUE,
  withClosing = TRUE
)
}
\arguments{
\item{name}{Cohort name listed as section title}

\item{json}{JSON cohort definition to be printed}

\item{obj}{Cohort object outputted from \code{CirceR} to be printed; can be \code{NULL}
in which case \code{json} is used.}

\item{withConcepts}{Boolean: Include concept lists in output?}

\item{withClosing}{Boolean: Add the output from \code{printCohortClose} to end?}
}
\description{
Outputs a cohort definition into human-readable \code{markdown}
}
