% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printConceptSet}
\alias{printConceptSet}
\title{Print concept set}
\usage{
printConceptSet(conceptSet, latexTableFontSize = 8)
}
\arguments{
\item{conceptSet}{JSON concept set definition to be printed}

\item{latexTableFontSize}{Font size to use if output will be converted to PDF via \code{latex}}
}
\description{
Outputs a concept set into human-readable \code{markdown}
}
