% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printExitCriteria}
\alias{printExitCriteria}
\title{Print exit criteria}
\usage{
printExitCriteria(obj)
}
\arguments{
\item{obj}{Cohort object outputted from \code{CirceR} to be printed
in which case \code{json} is used.}
}
\description{
Outputs exit criteria into human-readable \code{markdown}
}
