% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printInclusionCriteria}
\alias{printInclusionCriteria}
\title{Print inclusion criteria}
\usage{
printInclusionCriteria(obj, removeDescription = FALSE)
}
\arguments{
\item{obj}{Cohort object outputted from \code{CirceR} to be printed
in which case \code{json} is used.}

\item{removeDescription}{Currently not used (TODO fix)}
}
\description{
Outputs inclusion criteria into human-readable \code{markdown}
}
