% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendDatabase.R
\name{removeResultsFromDatabase}
\alias{removeResultsFromDatabase}
\title{Remove results from the database server.}
\usage{
removeResultsFromDatabase(
  connectionDetails,
  schema,
  databaseId,
  specifications
)
}
\arguments{
\item{connectionDetails}{Object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{schema}{Schema on the postgres server where the tables have been created.}

\item{databaseId}{Data sources to remove}

\item{specifications}{Specification of results tables}
}
\description{
Requires the results data model tables have been created using the \code{\link{createResultsDataModel}} function.

Set the POSTGRES_PATH environmental variable to the path to the folder containing the psql executable to enable
bulk upload (recommended).
}
