% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{runCohortMethod}
\alias{runCohortMethod}
\title{Run the cohort method package}
\usage{
runCohortMethod(
  outputFolder,
  indicationId = "class",
  databaseId,
  maxCores = 4,
  runSections,
  limitHois = NULL
)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/).
Do not use a folder on a network drive since this greatly impacts performance.}

\item{indicationId}{A string denoting the indicationId.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{maxCores}{How many parallel cores should be used? If more cores are made available this
can speed up the analyses.}
}
\description{
Run the cohort method package
}
\details{
Runs the cohort method package to produce propensity scores and outcome models.
}
