% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{unnumberAdditionalCriteria}
\alias{unnumberAdditionalCriteria}
\title{Find the subsections corresponding to additional criteria and unnumber them.}
\usage{
unnumberAdditionalCriteria(markdown)
}
\description{
Find the subsections corresponding to additional criteria and unnumber them.
}
