% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendDatabase.R
\name{grantPermissionOnServer}
\alias{grantPermissionOnServer}
\title{Grant read-permission on server}
\usage{
grantPermissionOnServer(
  connectionDetails,
  schema,
  user = "legendt2dm_readonly"
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{schema}{Schema name for read-permission}

\item{user}{User name for read-permission}
}
\description{
Grant read-permission on server
}
\details{
Only PostgreSQL servers are supported.
}
